requirements.txt
wget
flask
peewee
requests
psutil
pycrypto
